feat(xhttp): HTTP retry transport
This change implements an HTTP retry transport following HTTP semantics
as defined in https://datatracker.ietf.org/doc/html/rfc9110.

When the transport is chained to an HTTP client transport, it retries
requests when all the following conditions are met:
- request is retryable, either because the request method is idempotent
  or the request contains an Idempotency-Key header,
- response contains a retryable status code (with an eventual
  "Retry-After" header): 408, 413, 425, 429, 500, 502, 503 or 504.

Requests are retried with a configurable backoff policy, which defaults
to:
- an initial interval of 200 milliseconds,
- an interval multiplier of 1.5,
- a jitter factor of 0.2,
- a max interval of 30 seconds.

An optional xhttptrace.ClientTrace with a Retry hook can be added to the
context passed to the http.Request object to instrument/inspect the
retry logic.