fix(xtime): ensure TimeMilli serializes with trailing zeros
TimeMilli type is meant to breach the gap that exists in terms of time
serialization between languages like Javascript (millisecond precision)
and Golang (nanosecond precision).

The JSON encoding would result in a millisecond-precision ISO date time
with trailing zeros for fractional seconds, which is what this fix
implements as well.