feat(xerrors): implement Join which deprecates Append
Following introduction of errors.Join in 1.20, deprecate xerrors.Append
and replace by it by xerrors.Join.

xerrors.Join is reimplemented but mostly aligned with errors.Join to
appropriately take stack traces when necessary. Function can be used as
a drop-in replacement for errors.Join.

String serialization of joined errors differ from the Go SDK to keep
indentation so that a tree of joined errors can be better appreciated,
as opposed to the flat structure provided by errors.Join.